==> proteomics/0-create_study_design_tables.Rmd <==
---
title: "PTRC Exp 18 Study design tables"
output: html_document
---

```{r setup}
library(PNNL.DMS.utils)
library(stringr)
library(dplyr)
```



```{r}
path_to_global_study_design <- "study_design/global_study_design/"
path_to_phospho_study_design <- "study_design/phospho_study_design/"

dir.create(path_to_global_study_design)
dir.create(path_to_phospho_study_design)

global_data_package_num <- 4135
phospho_data_package_num <- 4136

```


Samples, references and fractions table must be created from study-specific metadata. 

See `r help(read_study_design)` for more information.


```{r samples}
library(readxl)

metadataTable <- read_excel("data/Exp18_metadata_01-14-2021.xlsx")

samples <- metadataTable %>% as.data.frame() %>%
  dplyr::select(Index, ReporterName = Channel, 
                PlexID = Plex, fullName) %>%
  mutate(QuantBlock = "1",
         ReporterAlias = as.character(Index),
         PlexID = as.character(PlexID),
         ReporterAlias = case_when(grepl("eference", fullName) ~ "ref",
                                   TRUE ~ ReporterAlias),
         ReporterName = case_when(ReporterName == "126C" ~ "126",       ## Here we make the channel names match the
                                  ReporterName == "134" ~ "134N",       ## format of the converter_ion conversion table.
                                  TRUE ~ ReporterName),
         MeasurementName = as.character(Index))

samples_p <- samples
samples_p[grepl("Pooled", samples_p$fullName), "MeasurementName"] <- c("Pooled 1", "Pooled 2")
samples_p[grepl("Pooled", samples_p$fullName), "ReporterAlias"] <- c("Pooled 1", "Pooled 2")
samples_p[samples_p$ReporterAlias == "ref" | is.na(samples_p$ReporterAlias), "MeasurementName"] <- NA

samples[grepl("Pooled", samples$MeasurementName), "MeasurementName"] <- NA
samples[samples$ReporterAlias == "ref" | is.na(samples$ReporterAlias), "MeasurementName"] <- NA

write.table(samples %>% select(-Index, -fullName), 
            paste0(path_to_global_study_design, "samples.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples_p %>% select(-Index, -fullName), 
            paste0(path_to_global_study_design, "samples_pooled.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples %>% select(-Index, -fullName), 
            paste0(path_to_phospho_study_design, "samples.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples_p %>% select(-Index, -fullName), 
            paste0(path_to_phospho_study_design, "samples_pooled.txt"),
            quote=F, sep="\t", row.names=F)

```


```{r references}
references <- samples %>%
  dplyr::select(PlexID, QuantBlock) %>%
  unique() %>% mutate(Reference = "ref")

write.table(references, paste0(path_to_global_study_design, "references.txt"),
            quote=F, sep="\t", row.names=F)

write.table(references, paste0(path_to_phospho_study_design, "references.txt"),
            quote=F, sep="\t", row.names=F)

```


```{r fractions}
## Global
job_records <- get_job_records_by_dataset_package(global_data_package_num)

fractions <- data.frame(Dataset = unique(job_records$Dataset)) %>%
  mutate(PlexID = sub("^PTRC_Ex18_Plex(.*)_G_.*$", "\\1", Dataset))

write.table(fractions, paste0(path_to_global_study_design, "fractions.txt"),
            quote=F, sep="\t", row.names=F)


## Phospho
job_records <- get_job_records_by_dataset_package(phospho_data_package_num)

fractions <- data.frame(Dataset = unique(job_records$Dataset)) %>%
  mutate(PlexID = sub("^PTRC_Ex18_Plex(.*)_P_.*$", "\\1", Dataset))

write.table(fractions, paste0(path_to_phospho_study_design, "fractions.txt"),
            quote=F, sep="\t", row.names=F)

```
