==> proteomics/3-combine_and_correct_datasets_global.Rmd <==
---
title: "Ex18 Combining datasets (Global)"
output: html_document
---




```{r include=FALSE}
library(MSnSet.utils)
library(dplyr)
library(ggplot2)
library(readxl)

source("../util/synapseUtil.R")
syn <- synapseLogin()

```


## Ex 18 data setup


```{r}
peptide_crosstab <- read.table("data/global_data/ptrc_ex18_crosstab_global_peptide_original.txt", sep = "\t")
peptide_df <- peptide_crosstab %>% select(Gene = accession, Peptide = peptide)
peptide_crosstab <- peptide_crosstab %>%
  select(-peptide, -accession, -pepSeq)
rownames(peptide_crosstab) <- peptide_df$Peptide
colnames(peptide_crosstab) <- sub("^X", "", colnames(peptide_crosstab))

meta <- read_excel("../proteomics/data/Exp18_metadata_01-14-2021.xlsx") %>%
  filter(!is.na(Index)) %>%
  mutate(Index = as.character(Index),
         Plex = paste0("Ex18_", as.character(Plex)),
         Sample = paste0("Ex18_Sample_", Index)) %>% as.data.frame()
rownames(meta) <- meta$Index
meta <- meta[colnames(peptide_crosstab), ]
colnames(peptide_crosstab) <- meta$Sample
rownames(meta) <- meta$Sample

m_ex18 <- MSnSet(exprs = as.matrix(peptide_crosstab), pData = meta)
plot_pca(m_ex18, "Plex")

```


### Missing data


```{r}
peptide_missing <- rowSums(is.na(peptide_crosstab))

hist(peptide_missing)

```

## Ex 12 data setup

```{r}
ex_12 <- read.table(syn$get("syn22136307")$path, sep = "\t", header = T)
ex_12_peptide_df <- ex_12 %>% select(Peptide, Gene) %>% unique()
ex_12 <- ex_12 %>%
  select(-Protein, -Gene) %>%
  unique()
ex_12_rownames <- ex_12$Peptide
ex_12 <- ex_12 %>% select(-Peptide)
rownames(ex_12) <- ex_12_rownames

## Plex 1 and 2 contain MOLM14, while Plex 3 and 4 contain MV411. So select only 1 and 2 here.
meta_ex12 <- read_excel("data/LabelingMetadata_exp12.xlsx", sheet = "Sheet1") %>%
  mutate(sample = paste0("Sample_", `Sample ID`),
         Sample = paste0("Ex12_", sample),
         Plex = paste0("Ex12_", as.character(Plex)),
         state = sub("(^[A-Z,a-z]+) .*$", "\\1", `Sample Info`),
         State = case_when(state == "parental" ~ "Parental",
                           TRUE ~ state),
         Treatment = case_when(State == "Parental" ~ "none",
                               TRUE ~ "G"),
         Ligand = case_when(State == "Parental" ~ "none",
                            TRUE ~ sub("^.*, ([A-Z,0-9]+)$", "\\1", `Sample Info`))) %>%
  head(40) %>% as.data.frame()
rownames(meta_ex12) <- meta_ex12$sample
meta_ex12 <- meta_ex12[colnames(ex_12), ]
colnames(ex_12) <- meta_ex12$Sample
rownames(meta_ex12) <- meta_ex12$Sample

peptide_crosstab_ex12 <- ex_12[, meta_ex12 %>% filter(Plex %in% c("Ex12_1", "Ex12_2")) %>% pull(Sample)]

m_ex12 <- MSnSet(exprs = as.matrix(peptide_crosstab_ex12), 
                 pData = meta_ex12[colnames(peptide_crosstab_ex12), ])

plot_pca(m_ex12, "Plex")

```

### Missing data

```{r}
peptide_ex12_missing <- rowSums(is.na(peptide_crosstab_ex12))
hist(peptide_ex12_missing)

```




Now we take the peptides which appear in each experiment and intersect to find a common
set of identified peptides.


```{r}
common_identified_peptides <- intersect(rownames(ex_12), rownames(peptide_crosstab))
peptide_df <- peptide_df %>%
  filter(Peptide %in% common_identified_peptides) %>%
  mutate(ID = paste(Peptide, Gene, sep = "_"))
ex_12_peptide_df <- ex_12_peptide_df %>%
  filter(Peptide %in% common_identified_peptides) %>%
  mutate(ID = paste(Peptide, Gene, sep = "_"))
ex_12_peptide_df <- ex_12_peptide_df %>% group_by(Peptide) %>% 
  mutate(total = n()) %>% ungroup()

multi_peptides <- ex_12_peptide_df %>% filter(total > 1) %>% pull(Peptide) %>% unique()
multi_peptide_IDs <- peptide_df %>% filter(Peptide %in% multi_peptides) %>% pull(ID) %>% unique()

## Settling on common fData
table(multi_peptide_IDs %in% ex_12_peptide_df$ID)
table(peptide_df$ID %in% ex_12_peptide_df$ID)
table(ex_12_peptide_df$ID %in% peptide_df$ID)

f_df = peptide_df %>% filter(ID %in% intersect(peptide_df$ID, ex_12_peptide_df$ID %>% unique()))
rownames(f_df) <- f_df$Peptide

peptide_crosstab <- peptide_crosstab[f_df$Peptide, ]
peptide_crosstab_ex12 <- peptide_crosstab_ex12[f_df$Peptide, ]

meta_combined <- rbind(meta %>% select(Sample, Plex, State, Treatment, Ligand), 
                       meta_ex12 %>% filter(Plex %in% c("Ex12_1", "Ex12_2")) %>% select(Sample, Plex, State, Treatment, Ligand)) %>%
  mutate(Exp = sub("_[0-9]$", "", Plex))

```


## Raw data stats


```{r}
## Raw data stats
# hist(apply(peptide_crosstab, 1, median, na.rm = T), main = "Ex 18 peptide medians")
# hist(apply(peptide_crosstab_ex12, 1, median, na.rm = T), main = "Ex 12 peptide medians")
# hist(apply(peptide_crosstab, 1, sd, na.rm = T), main = "Ex 18 peptide sds")
# hist(apply(peptide_crosstab_ex12, 1, sd, na.rm = T), main = "Ex 12 peptide sds")

# peptide_crosstab <- sweep(peptide_crosstab, 1, apply(peptide_crosstab, 1, median, na.rm = T), FUN = '-')
peptide_crosstab <- sweep(peptide_crosstab, 2, apply(peptide_crosstab, 2, median, na.rm = T), FUN = '-')
# peptide_crosstab_ex12 <- sweep(peptide_crosstab_ex12, 1, apply(peptide_crosstab_ex12, 1, median, na.rm = T), FUN = '-')
peptide_crosstab_ex12 <- sweep(peptide_crosstab_ex12, 2, apply(peptide_crosstab_ex12, 2, median, na.rm = T), FUN = '-')

```

The distribtuion in each dataset is varied, even though we have similar cell lines.
So, we will normalize using a single mean and standard deviation for each dataset, 
independent of peptides, batches, etc, to better compare the two datasets. 


```{r}
hist(as.matrix(peptide_crosstab), main = "Ex18 data distribution")
hist(as.matrix(peptide_crosstab_ex12), main = "Ex12 data distribution")

```


```{r}
ex18_median <- median(as.matrix(peptide_crosstab), na.rm = T)
ex18_sd <- sd(as.matrix(peptide_crosstab), na.rm = T)
ex12_median <- median(as.matrix(peptide_crosstab_ex12), na.rm = T)
ex12_sd <- sd(as.matrix(peptide_crosstab_ex12), na.rm = T)

crosstab_ex18 <- peptide_crosstab
crosstab_ex12 <- peptide_crosstab_ex12

crosstab_ex18_norm <- (peptide_crosstab - ex18_median)/ex18_sd
crosstab_ex12_norm <- (peptide_crosstab_ex12 - ex12_median)/ex12_sd

```


## Combining the data 


### Normalized first, combined then corrected.


```{r}
crosstab_combined_norm <- cbind(crosstab_ex18_norm, crosstab_ex12_norm)
crosstab_combined_norm <- sweep(crosstab_combined_norm, 1, 
                                apply(crosstab_combined_norm, 1, median, na.rm = T), FUN = '-')

## Missing data 
combined_missing <- rowSums(is.na(crosstab_combined_norm))
crosstab_combined_norm <- crosstab_combined_norm[combined_missing < 30, ]

m <- MSnSet(exprs = as.matrix(crosstab_combined_norm), pData = meta_combined, 
            fData = f_df[rownames(crosstab_combined_norm), ])
m_corrected <- correct_batch_effect_NA(m, batch_name = "Plex", par.prior = TRUE)
m_corrected_rollup <- rrollup(m_corrected, rollBy = "Gene", rollFun = '-')

plot_pca(m, "Plex")
# plot_pca(m, "Plex", components = c(3,4))
plot_pca(m_corrected, "Plex")
# plot_pca(m_corrected, "Plex", components = c(3,4))
plot_pca(m_corrected, "Treatment")
# plot_pca(m_corrected, "Treatment", components = c(3,4))
plot_pca(m_corrected_rollup, "Plex")
# plot_pca(m_corrected_rollup, "Plex")
plot_pca(m_corrected_rollup, "Treatment")
# plot_pca(m_corrected_rollup, "Treatment", components = c(3,4))

```


### Combined (not normalized) then corrected


```{r}
crosstab_combined <- cbind(crosstab_ex18, crosstab_ex12)
crosstab_combined <- sweep(crosstab_combined, 1, 
                           apply(crosstab_combined, 1, median, na.rm = T), FUN = '-')

## Missing data 
combined_missing <- rowSums(is.na(crosstab_combined))
crosstab_combined <- crosstab_combined[combined_missing < 30, ]

m <- MSnSet(exprs = as.matrix(crosstab_combined), pData = meta_combined, 
            fData = f_df[rownames(crosstab_combined), ])
m_corrected <- correct_batch_effect_NA(m, batch_name = "Plex", par.prior = TRUE)
m_corrected_rollup <- rrollup(m_corrected, rollBy = "Gene", rollFun = '-')

plot_pca(m, "Plex") + ggtitle("Combined dataset (Peptide) - uncorrected")
# plot_pca(m, "Plex", components = c(3,4))
plot_pca(m_corrected, "Plex") + ggtitle("Combined dataset (Peptide) - corrected")
# plot_pca(m_corrected, "Plex", components = c(3,4))
plot_pca(m_corrected, "Treatment") + ggtitle("Combined dataset (Peptide) - corrected")
# plot_pca(m_corrected, "Treatment", components = c(3,4))
plot_pca(m_corrected_rollup, "Plex") + ggtitle("Combined dataset (Protein) - corrected")
# plot_pca(m_corrected_rollup, "Plex")
plot_pca(m_corrected_rollup, "Treatment") + ggtitle("Combined dataset (Protein) - corrected")
# plot_pca(m_corrected_rollup, "Treatment", components = c(3,4))

## Using this one!
write.table(exprs(m_corrected_rollup), sep = "\t", file = "data/global_data/exp_12_18_combined_global_corrected.txt", quote = F)
write.table(pData(m_corrected_rollup), sep = "\t", file = "data/exp_12_18_combined_metadata.txt", quote = F)

```


### Corrected within each dataset first, then combined and corrected again.

NOT median centering each peptide before correction.

```{r}
mat <- crosstab_ex18
# mat <- sweep(mat, 1, apply(mat, 1, median, na.rm = T), FUN = '-')
n_missing <- rowSums(is.na(crosstab_combined))
mat <- mat[n_missing < ncol(mat)*0.5, ]

m_ex18 <- MSnSet(exprs = as.matrix(mat), pData = meta)


mat <- crosstab_ex12
# mat <- sweep(mat, 1, apply(mat, 1, median, na.rm = T), FUN = '-')
n_missing <- rowSums(is.na(crosstab_combined))
mat <- mat[n_missing < ncol(mat)*0.5, ]

m_ex12 <- MSnSet(exprs = as.matrix(mat), pData = meta_ex12 %>% filter(Plex %in% c("Ex12_1", "Ex12_2")))

m_ex18_corrected <- correct_batch_effect_NA(m_ex18, "Plex", par.prior = TRUE)
m_ex12_corrected <- correct_batch_effect_NA(m_ex12, "Plex", par.prior = TRUE)

# plot_pca(m_ex18_corrected, "Plex")
# plot_pca(m_ex12_corrected, "Plex")

common_peptides_corrected <- intersect(rownames(exprs(m_ex18)), rownames(exprs(m_ex12)))
mat_combined <- cbind(exprs(m_ex18_corrected)[common_peptides_corrected, ], exprs(m_ex12_corrected)[common_peptides_corrected, ])
mat_combined <- sweep(mat_combined, 1, 
                      apply(mat_combined, 1, median, na.rm = T), FUN = '-')

m <- MSnSet(exprs = mat_combined, pData = meta_combined, 
            fData = f_df[rownames(mat_combined), ])
m_corrected <- correct_batch_effect_NA(m, "Exp", par.prior = TRUE)
m_corrected_rollup <- rrollup(m_corrected, rollBy = "Gene", rollFun = '-')

plot_pca(m, "Plex")
# plot_pca(m, "Plex", components = c(3,4))
plot_pca(m_corrected, "Plex")
# plot_pca(m_corrected, "Plex", components = c(3,4))
plot_pca(m_corrected, "Treatment")
# plot_pca(m_corrected, "Treatment", components = c(3,4))
plot_pca(m_corrected_rollup, "Plex")
# plot_pca(m_corrected_rollup, "Plex")
plot_pca(m_corrected_rollup, "Treatment")
# plot_pca(m_corrected_rollup, "Treatment", components = c(3,4))

```

Here each peptide IS median centered

```{r}
mat <- crosstab_ex18
mat <- sweep(mat, 1, apply(mat, 1, median, na.rm = T), FUN = '-')
n_missing <- rowSums(is.na(crosstab_combined))
mat <- mat[n_missing < ncol(mat)*0.5, ]

m_ex18 <- MSnSet(exprs = as.matrix(mat), pData = meta)


mat <- crosstab_ex12
mat <- sweep(mat, 1, apply(mat, 1, median, na.rm = T), FUN = '-')
n_missing <- rowSums(is.na(crosstab_combined))
mat <- mat[n_missing < ncol(mat)*0.5, ]

m_ex12 <- MSnSet(exprs = as.matrix(mat), pData = meta_ex12 %>% filter(Plex %in% c("Ex12_1", "Ex12_2")))

m_ex18_corrected <- correct_batch_effect_NA(m_ex18, "Plex", par.prior = TRUE)
m_ex12_corrected <- correct_batch_effect_NA(m_ex12, "Plex", par.prior = TRUE)

# plot_pca(m_ex18_corrected, "Plex")
# plot_pca(m_ex12_corrected, "Plex")

common_peptides_corrected <- intersect(rownames(exprs(m_ex18)), rownames(exprs(m_ex12)))
mat_combined <- cbind(exprs(m_ex18_corrected)[common_peptides_corrected, ], exprs(m_ex12_corrected)[common_peptides_corrected, ])
mat_combined <- sweep(mat_combined, 1, 
                      apply(mat_combined, 1, median, na.rm = T), FUN = '-')

m <- MSnSet(exprs = mat_combined, pData = meta_combined, 
            fData = f_df[rownames(mat_combined), ])
m_corrected <- correct_batch_effect_NA(m, "Exp", par.prior = TRUE)
m_corrected_rollup <- rrollup(m_corrected, rollBy = "Gene", rollFun = '-')

plot_pca(m, "Plex")
# plot_pca(m, "Plex", components = c(3,4))
plot_pca(m_corrected, "Plex")
# plot_pca(m_corrected, "Plex", components = c(3,4))
plot_pca(m_corrected, "Treatment")
# plot_pca(m_corrected, "Treatment", components = c(3,4))
plot_pca(m_corrected_rollup, "Plex")
# plot_pca(m_corrected_rollup, "Plex")
plot_pca(m_corrected_rollup, "Treatment")
# plot_pca(m_corrected_rollup, "Treatment", components = c(3,4))
```


## Correcting at the Gene level

Ex 18


```{r}
library(MSnSet.utils)
library(dplyr)
library(ggplot2)
library(readxl)
source("../util/synapseUtil.R")
syn <- synapseLogin()

gene_crosstab_ex18 <- read.table("../proteomics/data/global_data/ptrc_ex18_crosstab_global_gene_original.txt")
colnames(gene_crosstab_ex18) <- sub("^X", "", colnames(gene_crosstab_ex18))
ex18_genes <- rownames(gene_crosstab_ex18)

meta <- read_excel("../proteomics/data/Exp18_metadata_01-14-2021.xlsx") %>%
  filter(!is.na(Index)) %>%
  mutate(Index = as.character(Index),
         Plex = paste0("Ex18_", as.character(Plex)),
         Sample = paste0("Ex18_Sample_", Index)) %>% as.data.frame()
rownames(meta) <- meta$Index
meta <- meta[colnames(gene_crosstab_ex18), ]
colnames(gene_crosstab_ex18) <- meta$Sample
rownames(meta) <- meta$Sample

m_ex18 <- MSnSet(exprs = as.matrix(gene_crosstab_ex18), pData = meta)
plot_pca(m_ex18, "Plex")


```



### Missing data


```{r}
gene_missing <- rowSums(is.na(gene_crosstab_ex18))
hist(gene_missing)

```

## Ex 12 data setup

```{r}
ex_12 <- read.table(syn$get("syn22136301")$path, sep = "\t", header = T)
ex_12 <- ex_12 %>%
  select(-Protein) %>%
  unique()
ex_12_gene_df <- ex_12 %>% select(Gene) %>% 
  group_by(Gene) %>%
  mutate(total = n()) %>% 
  ungroup()
ex12_genes <- ex_12_gene_df %>% filter(total == 1) %>% pull(Gene) %>% unique()
ex_12 <- ex_12 %>%
  filter(Gene %in% ex12_genes)
ex_12_rownames <- ex_12$Gene
ex_12 <- ex_12 %>% select(-Gene)
rownames(ex_12) <- ex_12_rownames

## Plex 1 and 2 contain MOLM14, while Plex 3 and 4 contain MV411. So select only 1 and 2 here.
meta_ex12 <- read_excel("data/LabelingMetadata_exp12.xlsx", sheet = "Sheet1") %>%
  mutate(sample = paste0("Sample_", `Sample ID`),
         Sample = paste0("Ex12_", sample),
         Plex = paste0("Ex12_", as.character(Plex)),
         state = sub("(^[A-Z,a-z]+) .*$", "\\1", `Sample Info`),
         State = case_when(state == "parental" ~ "Parental",
                           TRUE ~ state),
         Treatment = case_when(State == "Parental" ~ "none",
                               TRUE ~ "G"),
         Ligand = case_when(State == "Parental" ~ "none",
                            TRUE ~ sub("^.*, ([A-Z,0-9]+)$", "\\1", `Sample Info`))) %>%
  head(40) %>% as.data.frame()
rownames(meta_ex12) <- meta_ex12$sample
meta_ex12 <- meta_ex12[colnames(ex_12), ]
colnames(ex_12) <- meta_ex12$Sample
rownames(meta_ex12) <- meta_ex12$Sample

gene_crosstab_ex12 <- ex_12[, meta_ex12 %>% filter(Plex %in% c("Ex12_1", "Ex12_2")) %>% pull(Sample)]

m_ex12 <- MSnSet(exprs = as.matrix(gene_crosstab_ex12), 
                 pData = meta_ex12[colnames(gene_crosstab_ex12), ])

plot_pca(m_ex12, "Plex")
plot_pca(m_ex12, "Treatment")
plot_pca(m_ex12, "State")

```


### Missing data


```{r}
gene_ex12_missing <- rowSums(is.na(gene_crosstab_ex12))
hist(gene_ex12_missing)

```




Now we take the genes which appear in each experiment and intersect to find a common
set of identified genes


```{r}
common_identified_genes <- intersect(rownames(ex_12), rownames(gene_crosstab_ex18))

meta_combined <- rbind(meta %>% select(Sample, Plex, State, Treatment, Ligand), 
                       meta_ex12 %>% filter(Plex %in% c("Ex12_1", "Ex12_2")) %>% select(Sample, Plex, State, Treatment, Ligand)) %>%
  mutate(Exp = sub("_[0-9]$", "", Plex))

gene_crosstab_ex12 <- gene_crosstab_ex12[common_identified_genes, ]
gene_crosstab_ex18 <- gene_crosstab_ex18[common_identified_genes, ]

```


## Raw data stats


```{r}
gene_crosstab_ex12 <- sweep(gene_crosstab_ex12, 2, apply(gene_crosstab_ex12, 2, median, na.rm = T), FUN = '-')
gene_crosstab_ex18 <- sweep(gene_crosstab_ex18, 2, apply(gene_crosstab_ex18, 2, median, na.rm = T), FUN = '-')

```

The distribtuion in each dataset is varied, even though we have similar cell lines.
So, we will normalize using a single mean and standard deviation for each dataset, 
independent of peptides, batches, etc, to better compare the two datasets. 


```{r}
hist(as.matrix(gene_crosstab_ex18), main = "Ex18 data distribution")
hist(as.matrix(gene_crosstab_ex12), main = "Ex12 data distribution")

```


```{r}
ex18_median <- median(as.matrix(gene_crosstab_ex18), na.rm = T)
ex18_sd <- sd(as.matrix(gene_crosstab_ex18), na.rm = T)
ex12_median <- median(as.matrix(gene_crosstab_ex12), na.rm = T)
ex12_sd <- sd(as.matrix(gene_crosstab_ex12), na.rm = T)

crosstab_ex18 <- gene_crosstab_ex18
crosstab_ex12 <- gene_crosstab_ex12

crosstab_ex18_norm <- (gene_crosstab_ex18 - ex18_median)/ex18_sd
crosstab_ex12_norm <- (gene_crosstab_ex12 - ex12_median)/ex12_sd

```


## Combining the data 


### Normalized first, combined then corrected.


```{r}
crosstab_combined_norm <- cbind(crosstab_ex18_norm, crosstab_ex12_norm)
crosstab_combined_norm <- sweep(crosstab_combined_norm, 1, 
                                apply(crosstab_combined_norm, 1, median, na.rm = T), FUN = '-')

## Missing data 
combined_missing <- rowSums(is.na(crosstab_combined_norm))
crosstab_combined_norm <- crosstab_combined_norm[combined_missing < 30, ]

m <- MSnSet(exprs = as.matrix(crosstab_combined_norm), pData = meta_combined)
m_corrected <- correct_batch_effect_NA(m, batch_name = "Plex", par.prior = TRUE)

plot_pca(m, "Plex")
# plot_pca(m, "Plex", components = c(3,4))
plot_pca(m_corrected, "Plex")
# plot_pca(m_corrected, "Plex", components = c(3,4))
plot_pca(m_corrected, "Treatment")
# plot_pca(m_corrected, "Treatment", components = c(3,4))

```

```{r}
crosstab_combined <- cbind(crosstab_ex18, crosstab_ex12)
crosstab_combined <- sweep(crosstab_combined, 1, 
                           apply(crosstab_combined, 1, median, na.rm = T), FUN = '-')

## Missing data 
combined_missing <- rowSums(is.na(crosstab_combined))
crosstab_combined <- crosstab_combined[combined_missing < 30, ]

m <- MSnSet(exprs = as.matrix(crosstab_combined), pData = meta_combined)
m_corrected <- correct_batch_effect_NA(m, batch_name = "Plex", par.prior = TRUE)

plot_pca(m, "Plex")
# plot_pca(m, "Plex", components = c(3,4))
plot_pca(m_corrected, "Plex")
# plot_pca(m_corrected, "Plex", components = c(3,4))
plot_pca(m_corrected, "Treatment")
# plot_pca(m_corrected, "Treatment", components = c(3,4))

```
