==> proteomics/3-normalize_and_batch_correction_outdated.Rmd <==
---
title: "Ex18 Phospho proteomics"
output: html_document
---

Preliminary summary




```{r include=FALSE}
library(MSnSet.utils)
library(dplyr)
library(ggplot2)
library(readxl)

global_crosstab <- read.table("../proteomics/data/global_data/ptrc_ex18_crosstab_global_gene_original.txt")
phospho_crosstab <- read.table("../proteomics/data/phospho_data/ptrc_ex18_crosstab_phospho_siteID_original.txt")
phospho_crosstab_prot <- read.table("../proteomics/data/phospho_data/ptrc_ex18_crosstab_phospho_siteID_prot_original.txt", sep = "\t")

meta <- read_excel("../proteomics/data/Exp18_metadata_01-14-2021.xlsx") %>%
  filter(!is.na(Index)) %>%
  mutate(Index = as.character(Index),
         Plex = as.character(Plex)) %>% as.data.frame()
rownames(meta) <- meta$Index

colnames(global_crosstab) <- sub("^X", "", colnames(global_crosstab))
colnames(phospho_crosstab) <- sub("^X", "", colnames(phospho_crosstab))
colnames(phospho_crosstab_prot) <- sub("^data.", "", colnames(phospho_crosstab_prot))

## data only
phospho_prot_mat <- phospho_crosstab_prot[, intersect(as.character(1:47), colnames(phospho_crosstab_prot))]


```


Missing data


```{r}
global_missing <- rowSums(is.na(global_crosstab))
phospho_missing <- rowSums(is.na(phospho_crosstab))
phospho_prot_missing <- rowSums(is.na(phospho_prot_mat))

hist(global_missing)
hist(phospho_missing)
hist(phospho_prot_missing)

## 50 % missingness filter, lose 178 global features and 27205 phosphosites
global_crosstab <- global_crosstab[global_missing < ncol(global_crosstab)/2, ]
phospho_crosstab <- phospho_crosstab[phospho_missing < ncol(phospho_crosstab)/2, ]
phospho_prot_mat <- phospho_prot_mat[phospho_prot_missing < ncol(phospho_crosstab)/2, ]

```


Raw data stats.


```{r}
## row medians
hist(apply(global_crosstab, 1, median, na.rm = T))
hist(apply(phospho_crosstab, 1, median, na.rm = T))
hist(apply(phospho_prot_mat, 1, median, na.rm = T))

global_row_medians <- apply(global_crosstab, 1, median, na.rm = T)
phospho_row_medians <- apply(phospho_crosstab, 1, median, na.rm = T)
phospho_prot_row_medians <- apply(phospho_prot_mat, 1, median, na.rm = T)

global_crosstab <- sweep(global_crosstab, 1, global_row_medians, FUN = '-')
phospho_crosstab <- sweep(phospho_crosstab, 1, phospho_row_medians, FUN = '-')
phospho_prot_mat <- sweep(phospho_prot_mat, 1, phospho_prot_row_medians, FUN = '-')

```



```{r}
table(colnames(global_crosstab) == colnames(phospho_crosstab))
table(colnames(global_crosstab) == colnames(phospho_prot_mat))

## sample medians
hist(apply(global_crosstab, 2, median, na.rm = T))
hist(apply(phospho_crosstab, 2, median, na.rm = T))
hist(apply(phospho_prot_mat, 2, median, na.rm = T))

## Using global coefficients to normalize phospho.
global_sample_coef <- apply(global_crosstab, 2, median, na.rm = T)

global_crosstab <- sweep(global_crosstab, 2, global_sample_coef, FUN = '-')
phospho_crosstab <- sweep(phospho_crosstab, 2, global_sample_coef, FUN = '-')
phospho_prot_mat <- sweep(phospho_prot_mat, 2, global_sample_coef, FUN = '-')

```



```{r}
m_global <- MSnSet(exprs = global_crosstab %>% as.matrix(), 
                   pData = meta[colnames(global_crosstab), ])

m_phospho <- MSnSet(exprs = phospho_crosstab %>% as.matrix(), 
                    pData = meta[colnames(phospho_crosstab), ])

m_phospho_prot <- MSnSet(exprs = phospho_prot_mat %>% as.matrix(), 
                         pData = meta[colnames(phospho_prot_mat), ])

```


Both datasets show strong batch effect


```{r}
plot_pca(m_global, phenotype = "Plex") + ggtitle("Global PCA - All features by Plex")
plot_pca(m_global, phenotype = "Treatment") + ggtitle("Global PCA - All features by Plex")
plot_pca(m_phospho, phenotype = "Plex") + ggtitle("Phospho PCA - All features by Plex")

# plot_pca(m_phospho_prot, phenotype = "Plex") + ggtitle("Phospho prot PCA - All features by Plex")

```


```{r}
m_global_corrected <- correct_batch_effect_NA(m_global, "Plex", par.prior = T)
m_phospho_corrected <- correct_batch_effect_NA(m_phospho, "Plex", par.prior = T)

# m_phospho_prot_corrected <- correct_batch_effect_NA(m_phospho_prot, "Plex", par.prior = T)

```



```{r}
plot_pca(m_global_corrected, phenotype = "Plex") + ggtitle("Global PCA - All features by Plex, Corrected")
plot_pca(m_global_corrected, phenotype = "Treatment") + ggtitle("Global PCA - All features by Plex, Corrected")
plot_pca(m_phospho_corrected, phenotype = "Plex") + ggtitle("Phospho PCA - All features by Plex, Corrected")
plot_pca(m_phospho_corrected, phenotype = "Treatment") + ggtitle("Phospho PCA - All features by Plex, Corrected")

# plot_pca(m_phospho_prot_corrected, phenotype = "Plex") + ggtitle("Phospho prot PCA - All features by Plex, Corrected")

```

Push to synapse.

```{r}
write.table(exprs(m_global_corrected), 
            file = "data/global_data/ptrc_ex18_crosstab_global_gene_corrected.txt",
            quote=F, sep="\t")

write.table(exprs(m_phospho_corrected), 
            file = "data/phospho_data/ptrc_ex18_crosstab_phospho_SiteID_corrected.txt",
            quote=F, sep="\t")


phospho_prot_corrected <- exprs(m_phospho_prot_corrected) %>%
  as.data.frame() %>%
  mutate(feature = rownames(.))
mapping_df <- phospho_crosstab_prot %>%
  mutate(feature = rownames(.)) %>%
  select(feature, Peptide_kstar, Protein_kstar, SiteID)

phospho_prot_corrected <- merge(phospho_prot_corrected, mapping_df, by = "feature")
rownames(phospho_prot_corrected) <- phospho_prot_corrected$feature
phospho_prot_corrected <- phospho_prot_corrected %>% select(-feature) %>%
  select(SiteID, Peptide_kstar, Protein_kstar, everything())

colnames(phospho_prot_corrected) <- sub("(^[0-9].*$)", "data:\\1", colnames(phospho_prot_corrected))
write.table(phospho_prot_corrected, 
            file = "data/phospho_data/ptrc_ex18_crosstab_phospho_SiteID_prot_corrected.txt",
            sep="\t")

```


```{r}
plex_global <- limma_gen(m_global, "~Plex", "Plex")
plex_phospho <- limma_gen(m_phospho, "~Plex", "Plex")

plex_global_corrected <- limma_gen(m_global_corrected, "~Plex", "Plex")
plex_phospho_corrected <- limma_gen(m_phospho_corrected, "~Plex", "Plex")


```



```{r}
hist(plex_global$P.Value)
hist(plex_global_corrected$P.Value)

hist(plex_phospho$P.Value)
hist(plex_phospho_corrected$P.Value)

```





```{r}
library(readxl)
## Finalizing KSTAR input
crosstab <- read.table(file = "data/phospho_data/ptrc_ex18_crosstab_phospho_SiteID_prot_corrected.txt",
            sep="\t")
colnames(crosstab) <- sub("data.", "", colnames(crosstab))
meta <- read_excel("../proteomics/data/Exp18_metadata_01-14-2021.xlsx") %>%
  filter(!is.na(Index)) %>%
  mutate(Index = as.character(Index),
         Plex = as.character(Plex),
         comp1_ = paste(State, Treatment, sep = "_and_")) %>% as.data.frame()
rownames(meta) <- meta$Index

crosstab_kstar <- crosstab %>%
  select(SiteID, Peptide_kstar, Protein_kstar)

all_groups <- c("Early_and_G", "Late_and_G", "Early_and_GD", "Late_and_GD", 
                "Early_and_GV", "Late_and_GV", "Early_and_GVD", "Late_and_GVD")
for (comp1 in all_groups){
  samples <- meta %>% filter(comp1_ == comp1) %>%
    pull(Index)
  parental_samples <- meta %>% filter(comp1_ == "Parental_and_none") %>% 
    pull(Index)
  mat <- crosstab[, samples]
  crosstab_kstar[paste0("data:", comp1, "_logFC")] <- apply(crosstab[, samples], 1, mean, na.rm = T) - apply(crosstab[, parental_samples], 1, mean, na.rm = T)
}

data_folder = "data/phospho_data/"
crosstab_prefix = "ptrc_ex18"
## showing logFC vs Parental and none for each treatment.
write.table(crosstab_kstar,
            file=file.path(data_folder, paste(crosstab_prefix, "crosstab_phospho_treatments_logFC.txt", sep="_")),
            sep="\t", quote = F)

```
