==> proteomics/5-push_to_synapse.Rmd <==
---
title: "PTRC Ex18 Push to synapse"
output: html_document
---

```{r login}
library(amlresistancenetworks)
library(reticulate)

syn <- synapseLogin()

```


```{r upload}
synapse_id_map <- c("syn43185649" = "data/global_data/",
                    "syn43185649" = "data/phospho_data/")

for (i in seq_along(synapse_id_map)) {
  crosstabs <- list.files(path = synapse_id_map[i],
                          pattern = "crosstab.*.txt",
                          full.names = T)
  lapply(crosstabs, synapseStore,
         parentId = names(synapse_id_map)[i])
}

```


## Exp 12 + exp 18


```{r}
metadata <- read.table("../proteomics/data/exp_12_18_combined_metadata.txt", sep = "\t")

crosstab <- read.table("../proteomics/data/global_data/exp_12_18_combined_global_corrected.txt", sep = "\t")
crosstab_long <- crosstab %>%
  mutate(Gene = rownames(.)) %>%
  tidyr::pivot_longer(-Gene, names_to = "Sample", values_to = "value") %>%
  merge(metadata, by = "Sample")

write.table(crosstab_long, "../proteomics/data/global_data/exp_12_18_combined_global_corrected_long.txt", quote = F, sep = "\t")
synapseStore(path = "../proteomics/data/global_data/exp_12_18_combined_global_corrected_long.txt", parentId = "syn43185649")

crosstab <- read.table("../proteomics/data/phospho_data/exp_12_18_combined_phospho_corrected.txt", sep = "\t")
crosstab_long <- crosstab %>%
  mutate(SiteID = rownames(.)) %>%
  tidyr::pivot_longer(-SiteID, names_to = "Sample", values_to = "value") %>%
  merge(metadata, by = "Sample")

write.table(crosstab_long, "../proteomics/data/phospho_data/exp_12_18_combined_phospho_corrected_long.txt", quote = F, sep = "\t")
synapseStore(path = "../proteomics/data/phospho_data/exp_12_18_combined_phospho_corrected_long.txt", parentId = "syn43185649")


```
